+++ plugins/catalog-backend-module-ldap/src/ldap/client.ts
@@ -18,7 +18,7 @@ import { ForwardedError } from '@backstage/errors';
 import ldap, { Client, SearchEntry, SearchOptions } from 'ldapjs';
 import { cloneDeep } from 'lodash';
 import { Logger } from 'winston';
-import { BindConfig } from './config';
+import { BindConfig, TLSConfig } from './config';
 import { errorString } from './util';
 import {
   ActiveDirectoryVendor,
@@ -40,8 +40,12 @@ export class LdapClient {
     logger: Logger,
     target: string,
     bind?: BindConfig,
+    tls?: TLSConfig,
   ): Promise<LdapClient> {
-    const client = ldap.createClient({ url: target });
+    const client = ldap.createClient({
+      url: target,
+      tlsOptions: tls,
+    });
 
     // We want to have a catch-all error handler at the top, since the default
     // behavior of the client is to blow up the entire process when it fails,