+++ plugins/tech-insights-backend/src/service/persistence/TechInsightsDatabase.test.ts
@@ -115,6 +115,56 @@ const additionalFacts = [
   },
 ];
 
+const sameFactsDiffDateSchema = {
+  id: 'same-fact-diff-date-test',
+  version: '0.0.1-test',
+  entityFilter: JSON.stringify([{ kind: 'service' }]),
+  schema: JSON.stringify({
+    testStringFact: {
+      type: 'string',
+      description: 'Test fact with a string type',
+    },
+  }),
+};
+
+const sameFactsDiffDateNow = DateTime.now().toISO();
+const sameFactsDiffDateNearFuture = DateTime.now()
+  .plus(Duration.fromMillis(555))
+  .toISO();
+const sameFactsDiffDateFuture = DateTime.now()
+  .plus(Duration.fromMillis(1000))
+  .toISO();
+
+const multipleSameFacts = [
+  {
+    timestamp: sameFactsDiffDateNow,
+    id: sameFactsDiffDateSchema.id,
+    version: '0.0.1-test',
+    entity: 'a:a/a',
+    facts: JSON.stringify({
+      testNumberFact: 1,
+    }),
+  },
+  {
+    timestamp: sameFactsDiffDateNearFuture,
+    id: sameFactsDiffDateSchema.id,
+    version: '0.0.1-test',
+    entity: 'a:a/a',
+    facts: JSON.stringify({
+      testNumberFact: 2,
+    }),
+  },
+  {
+    timestamp: sameFactsDiffDateFuture,
+    id: 'multiple-same-facts',
+    version: '0.0.1-test',
+    entity: 'a:a/a',
+    facts: JSON.stringify({
+      testNumberFact: 3,
+    }),
+  },
+];
+
 describe('Tech Insights database', () => {
   const databases = TestDatabases.create();
   let store: TechInsightsStore;
@@ -213,6 +263,33 @@ describe('Tech Insights database', () => {
       'a:a/a',
     );
     expect(returnedFact['test-fact']).toMatchObject(baseAssertionFact);
+  });
+
+  it('should return latest fact with multiple entries', async () => {
+    await testDbClient.batchInsert('fact_schemas', [sameFactsDiffDateSchema]);
+    await testDbClient.batchInsert(
+      'facts',
+      multipleSameFacts.map(fact => ({
+        ...fact,
+        id: sameFactsDiffDateSchema.id,
+      })),
+    );
+
+    const returnedFacts = await store.getLatestFactsByIds(
+      ['test-fact', sameFactsDiffDateSchema.id],
+      'a:a/a',
+    );
+
+    expect(returnedFacts['test-fact']).toMatchObject({
+      ...baseAssertionFact,
+    });
+
+    expect(returnedFacts[sameFactsDiffDateSchema.id]).toMatchObject({
+      ...baseAssertionFact,
+      id: sameFactsDiffDateSchema.id,
+      timestamp: DateTime.fromISO(sameFactsDiffDateFuture),
+      facts: { testNumberFact: 3 },
+    });
   });
 
   it('should return latest facts for multiple ids', async () => {