+++ plugins/circleci/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-circleci
 
+## 0.3.6-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.1.1-next.0
+  - @backstage/core-components@0.9.5-next.0
+
 ## 0.3.5
 
 ### Patch Changes