+++ plugins/gcp-projects/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-gcp-projects
 
+## 0.3.25-next.0
+
+### Patch Changes
+
+- 6968b65ba1: Updated dependency `@react-hookz/web` to `^14.0.0`.
+- Updated dependencies
+  - @backstage/core-components@0.9.5-next.0
+
 ## 0.3.24
 
 ### Patch Changes