+++ plugins/github-actions/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-github-actions
 
+## 0.5.6-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.1.1-next.0
+  - @backstage/core-components@0.9.5-next.0
+  - @backstage/integration@1.2.1-next.0
+
 ## 0.5.5
 
 ### Patch Changes