+++ plugins/github-deployments/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-github-deployments
 
+## 0.1.37-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.1.1-next.0
+  - @backstage/core-components@0.9.5-next.0
+  - @backstage/integration@1.2.1-next.0
+  - @backstage/integration-react@1.1.1-next.0
+
 ## 0.1.36
 
 ### Patch Changes