+++ plugins/graphql-backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-graphql-backend
 
+## 0.1.23-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.13.6-next.0
+
 ## 0.1.22
 
 ### Patch Changes