+++ plugins/periskop/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-periskop
 
+## 0.1.4-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.1.1-next.0
+  - @backstage/core-components@0.9.5-next.0
+
 ## 0.1.3
 
 ### Patch Changes