+++ plugins/stack-overflow/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-stack-overflow
 
+## 0.1.2-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.9.5-next.0
+  - @backstage/plugin-home@0.4.22-next.0
+
 ## 0.1.1
 
 ### Patch Changes