+++ plugins/techdocs-addons-test-utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-techdocs-addons-test-utils
 
+## 1.0.1-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog@1.2.1-next.0
+  - @backstage/core-components@0.9.5-next.0
+  - @backstage/plugin-techdocs@1.1.2-next.0
+  - @backstage/plugin-techdocs-react@1.0.1-next.0
+  - @backstage/integration-react@1.1.1-next.0
+
 ## 1.0.0
 
 ### Major Changes